🚧[Refactor] 프로젝트 분리
