+++ server/src/constant/statusCode.js
@@ -1,0 +1,6 @@
+module.exports = {
+  OK: 200,
+  BAD_REQUEST: 400,
+  UNAUTHORIZED: 401,
+  INTERNAL_SERVER_ERROR: 500,
+};
